ai-service/requirements.txt
deepface==0.0.87
tensorflow==2.10.1
opencv-python-headless
numpy==1.23.5
pillow
fastapi
uvicorn
python-multipart
gunicorn
mtcnn